Split into parts.

+++ src/server/args.rs
@@ -26,10 +26,10 @@ impl From<LogLevel> for Level {
 #[derive(Parser, Debug, Serialize)]
 #[command(version, about, long_about = None)]
 pub struct Args {
-    #[arg(short, long, default_value = "0.0.0.0")]
+    #[arg(short, long, default_value = "localhost")]
     pub address: String,
     #[arg(short, long, default_value_t = 7654)]
-    pub port: u32,
+    pub port: u16,
     #[arg(long, default_value = "urocyon.database.sqlite3")]
     pub db_path: String,
     #[arg(long, default_value = "urocyon.log.ndjson")]
